% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.tblr}
\alias{print.tblr}
\title{Print a \strong{tblr} table object}
\usage{
\method{print}{tblr}(x, ...)
}
\arguments{
\item{x}{A \code{tblr} table object.}

\item{...}{Any additional parameters.}
}
\value{
A character string containing LaTeX markup.
}
\description{
This function uses \code{tblr_as_latex} to generate LaTeX markup from the \code{tblr} table object and prints it to the console.
}
