% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_boxhead_property.R
\name{set_column_labels}
\alias{set_column_labels}
\title{Set column labels}
\usage{
set_column_labels(x, ...)
}
\arguments{
\item{x}{A \code{tblr} table object.}

\item{...}{Expressions for the assignment of column labels. Expressions can be named arguments of the form \verb{<column name> = <label>} or make use of tidy-select. When using tidy-select, expressions have to be two-sided formulas (i.e., of the form \verb{<LHS> ~ <RHS>}). In this case, the left-hand side is used to identify selections of columns and the right-hand side to specify the column label. Note that the column labels you specify are assumed to be raw LaTeX strings; characters that have a special meaning in LaTeX will not be escaped.}
}
\value{
A \code{tblr} table object.
}
\description{
Define column labels (headers). By default, column labels are derived from the column names of the underlying dataframe. However, unlike those column names, column labels do not have to be unique. Note that column labels are only used for rendering the table; other functions still use the original column names to refer to columns of the underlying dataframe.
}
\examples{
library(dplyr)
df <- tibble(
  continent = c("Europe", "Asia", "Asia", "Europe", "Asia"),
  country = c("Germany","China", "Afghanistan","France","Taiwan"),
  value = c(0.17, 0.23, 11.3, 17, 2.4)
)

tblr(df) |>
  set_column_labels(value = "\\\\textbf{Value}", ends_with("y") ~ "COUNTRY")
}
