% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_column_spanner.R
\name{set_column_spanner}
\alias{set_column_spanner}
\title{Set column spanners}
\usage{
set_column_spanner(x, ...)
}
\arguments{
\item{x}{A \code{tblr} table object.}

\item{...}{Expressions for the assignment of column spanner labels. Expressions have to be two-sided formulas (i.e., of the form \verb{<LHS> ~ <RHS>}). The left-hand side is used to identify selections of columns via tidy-select and the right-hand side to specify the column spanner label. Note that the column spanner labels you provide are assumed to be raw LaTeX strings; characters that have a special meaning in LaTeX will not be escaped.}
}
\value{
A \code{tblr} table object.
}
\description{
Set column spanner labels. Column spanner labels are labels that appear above column labels (headers) and span several of them. Call this function once for each spanner row in the table, starting with the one at the bottom.
}
\examples{
library(dplyr)
df <- tibble(
  continent = c("Europe", "Asia", "Asia", "Europe", "Asia"),
  country = c("Germany","China", "Afghanistan","France","Taiwan"),
  value = c(0.17, 0.23, 11.3, 17, 2.4)
)

tblr(df) |>
  set_column_spanner(!value ~ "Region")
}
