% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_list_attribute.R
\name{set_options}
\alias{set_options}
\title{Set table options}
\usage{
set_options(x, ...)
}
\arguments{
\item{x}{A \code{tblr} table object.}

\item{...}{Named arguments to be spliced into key-value pairs within the mandatory argument of a \code{tblr} environment.}
}
\value{
A \code{tblr} table object.
}
\description{
This function gives you low-level access to the optional argument of a \code{tblr} environment, also called "outer specifications". For example, for types "float" and "break" the optional argument can contain the caption, a label, and notes.
}
\details{
The optional argument is a list of key-value pairs. Use named arguments to specify them with this function. Both the names and values you provide are assumed to be raw LaTeX strings; characters that have a special meaning in LaTeX will not be escaped. Each value is automatically wrapped in curly braces.

Shortcut functions targeting specific options such as the caption you specify in the call to \code{tblr()} take precedence over the specifications you make here.
}
