% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_list_attribute.R
\name{set_source_notes}
\alias{set_source_notes}
\title{Add source notes}
\usage{
set_source_notes(x, ...)
}
\arguments{
\item{x}{A \code{tblr} table object of type "float" or "break".}

\item{...}{Named arguments to add notes under a table. Names provide a label for each note; the argument value is the actual note itself.}
}
\value{
A \code{tblr} table object.
}
\description{
Add table notes to your table. Only \code{tblr} table objects of type "float" or "break" support table notes.
}
\examples{
library(dplyr)
df <- tibble(
  continent = c("Europe", "Asia", "Asia", "Europe", "Asia"),
  country = c("Germany","China", "Afghanistan","France","Taiwan"),
  value = c(0.17, 0.23, 11.3, 17, 2.4)
)

x <- tblr(df, type = "float", caption = "A table with notes")

x |> set_source_notes(
 Source = "Built-in R dataset",
 Note = "This dataset comes with every R installation and is available without explicitly loading it."
)
}
