% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabularray-package.R
\docType{package}
\name{tabularray-package}
\alias{tabularray}
\alias{tabularray-package}
\title{tabularray: Typeset LaTeX Tables using Tabularray}
\description{
Typeset R objects such as dataframes in LaTeX using 'tabularray'. 'tabularray' is a LaTeX package developed by Jianrui Lyu that provides a modern and unified alternative to the established space of table-generating packages in LaTeX. This implementation in R is inspired by the R package 'gt'. That means we construct a table iteratively, adding formatting by chaining functions together. Unlike 'gt', however, this package focuses on LaTeX output. It intends to maximize access the functionality of the (LaTeX) 'tabularray' package while offering convenience functions for the most common formatting tasks.
}
\author{
\strong{Maintainer}: Felix Turbanisch \email{felix.turbanisch@uni-goettingen.de} (\href{https://orcid.org/0000-0003-2279-1686}{ORCID})

}
\keyword{internal}
