% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tblr.R
\name{tblr}
\alias{tblr}
\title{Create a \strong{tblr} table object}
\usage{
tblr(df, type = "simple", caption = NULL)
}
\arguments{
\item{df}{The \code{tibble} or \code{data.frame} to be typeset as a table.}

\item{type}{The type of LaTeX table environment. This can be either "simple" (the default), "float", or "break".
\itemize{
\item "simple" produces an unnumbered and untitled table intended for simple tabular matter to appear in the run of the text (see CMOS 3.51).
\item "float" produces a floating numbered table that accommodates both a title and table notes, similar to a \code{threeparttable} in LaTeX.
\item "break" produces a numbered table that can be broken across pages instead of floating. It accommodates both a title and table notes, similar to a \code{longtable} in LaTeX.
}}

\item{caption}{Title of the table. Needs to be specified unless \code{type} is "simple". In LaTeX output, this title appears under the "caption" keyword.}
}
\value{
A \code{data.frame} of class \code{tblr}.
}
\description{
This function can be used to initialize a dataframe as a table object of class \code{tblr}. Usually, this is just the starting point: Use the pipe and chain together functions to incrementally add styling and additional parts (such as notes) to your table.
}
\details{
\code{tblr()} preserves the dataframe it is called upon so you can continue to use functions that operate on dataframes. This function only adds the class \code{tblr} to the object and saves some metadata about the table (such as column types) as attributes. This metadata determines how the table is rendered at the end of a workflow. If you modify columns after the call to \code{tblr()}, you risk breaking this link.

For example, characters that have a special meaning in LaTeX are escaped only in those columns that were originally text-like (character strings or factors). If you apply a formatting function to convert a numeric column to a character vector between your call to \code{tblr()} and rendering, you are required to take care of escaping special characters yourself.
}
\examples{
tblr(mtcars)
}
