% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tblr_as_latex.R
\name{tblr_as_latex}
\alias{tblr_as_latex}
\title{Output a \strong{tblr} object as LaTeX}
\usage{
tblr_as_latex(x)
}
\arguments{
\item{x}{A \code{tblr} table object.}
}
\value{
A character string containing LaTeX markup of class \code{knit_asis}.
}
\description{
Generate LaTeX markup to typeset a \code{tblr} object. The output is a character string of class \code{knit_asis} that is automatically included verbatim in R Markdown documents.
}
\details{
Make sure to load the necessary LaTeX packages in your document. In a Quarto document's YAML metadata, please include
\preformatted{
format:
  pdf:
    include-in-header: tabularray-packages.sty
}

The \code{tabularray_packages.sty} file would then contain the dependencies listed below:

\preformatted{
\usepackage{tabularray}
\UseTblrLibrary{booktabs}
}

You do not need to modify any chunk options. \code{knitr} will automatically embed the LaTeX markup verbatim.
}
\examples{
x <- tblr(mtcars)
tblr_as_latex(x)
}
